+++ spotbugs-tests/src/test/java/edu/umd/cs/findbugs/sarif/SarifBugReporterTest.java
@@ -445,6 +445,24 @@ class SarifBugReporterTest {
         }
     }
 
+    @Test
+    void testInvocationSuccessfulOnBugPresence() {
+        BugPattern bugPattern = new BugPattern("TYPE", "abbrev", "category", false, "shortDescription",
+                "longDescription", "detailText", "https://example.com/help.html", 0);
+        DetectorFactoryCollection.instance().registerBugPattern(bugPattern);
+
+        reporter.reportBug(new BugInstance(bugPattern.getType(), bugPattern.getPriorityAdjustment()).addInt(10).addClass("the/target/Class"));
+        reporter.finish();
+
+        String json = writer.toString();
+        JsonObject jsonObject = new Gson().fromJson(json, JsonObject.class);
+        JsonObject run = jsonObject.getAsJsonArray("runs").get(0).getAsJsonObject();
+        JsonObject invocation = run.getAsJsonArray("invocations").get(0).getAsJsonObject();
+
+        assertThat(invocation.get("exitCode").getAsInt(), is(1));
+        assertThat(invocation.get("exitCodeDescription").getAsString(), is("BUGS FOUND"));
+        assertThat(invocation.get("executionSuccessful").getAsBoolean(), is(true));
+    }
 
     Optional<String> takeFirstKey(JsonObject object) {
         return object.keySet().stream().findFirst();